Move cmd#readSource to loader#ReadSource
