+++ loader/readsource_test.go
@@ -1,4 +1,4 @@
-package cmd
+package loader
 
 import (
 	"bytes"
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/loadimpact/k6/lib/fsext"
-	"github.com/loadimpact/k6/loader"
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/require"
@@ -22,7 +21,7 @@ func (e errorReader) Read(_ []byte) (int, error) {
 var _ io.Reader = errorReader("")
 
 func TestReadSourceSTDINError(t *testing.T) {
-	_, err := readSource("-", "", nil, errorReader("1234"))
+	_, err := ReadSource("-", "", nil, errorReader("1234"))
 	require.Error(t, err)
 	require.Equal(t, "1234", err.Error())
 }
@@ -31,10 +30,10 @@ func TestReadSourceSTDINCache(t *testing.T) {
 	var data = []byte(`test contents`)
 	var r = bytes.NewReader(data)
 	var fs = afero.NewMemMapFs()
-	sourceData, err := readSource("-", "/path/to/pwd",
+	sourceData, err := ReadSource("-", "/path/to/pwd",
 		map[string]afero.Fs{"file": fsext.NewCacheOnReadFs(nil, fs, 0)}, r)
 	require.NoError(t, err)
-	require.Equal(t, &loader.SourceData{
+	require.Equal(t, &SourceData{
 		URL:  &url.URL{Scheme: "file", Path: "/-"},
 		Data: data}, sourceData)
 	fileData, err := afero.ReadFile(fs, "/-")
@@ -46,9 +45,9 @@ func TestReadSourceRelative(t *testing.T) {
 	var data = []byte(`test contents`)
 	var fs = afero.NewMemMapFs()
 	require.NoError(t, afero.WriteFile(fs, "/path/to/somewhere/script.js", data, 0644))
-	sourceData, err := readSource("../somewhere/script.js", "/path/to/pwd", map[string]afero.Fs{"file": fs}, nil)
+	sourceData, err := ReadSource("../somewhere/script.js", "/path/to/pwd", map[string]afero.Fs{"file": fs}, nil)
 	require.NoError(t, err)
-	require.Equal(t, &loader.SourceData{
+	require.Equal(t, &SourceData{
 		URL:  &url.URL{Scheme: "file", Path: "/path/to/somewhere/script.js"},
 		Data: data}, sourceData)
 }
@@ -59,9 +58,9 @@ func TestReadSourceAbsolute(t *testing.T) {
 	var fs = afero.NewMemMapFs()
 	require.NoError(t, afero.WriteFile(fs, "/a/b", data, 0644))
 	require.NoError(t, afero.WriteFile(fs, "/c/a/b", []byte("wrong"), 0644))
-	sourceData, err := readSource("/a/b", "/c", map[string]afero.Fs{"file": fs}, r)
+	sourceData, err := ReadSource("/a/b", "/c", map[string]afero.Fs{"file": fs}, r)
 	require.NoError(t, err)
-	require.Equal(t, &loader.SourceData{
+	require.Equal(t, &SourceData{
 		URL:  &url.URL{Scheme: "file", Path: "/a/b"},
 		Data: data}, sourceData)
 }
@@ -70,10 +69,10 @@ func TestReadSourceHttps(t *testing.T) {
 	var data = []byte(`test contents`)
 	var fs = afero.NewMemMapFs()
 	require.NoError(t, afero.WriteFile(fs, "/github.com/something", data, 0644))
-	sourceData, err := readSource("https://github.com/something", "/c",
+	sourceData, err := ReadSource("https://github.com/something", "/c",
 		map[string]afero.Fs{"file": afero.NewMemMapFs(), "https": fs}, nil)
 	require.NoError(t, err)
-	require.Equal(t, &loader.SourceData{
+	require.Equal(t, &SourceData{
 		URL:  &url.URL{Scheme: "https", Host: "github.com", Path: "/something"},
 		Data: data}, sourceData)
 }
@@ -82,7 +81,7 @@ func TestReadSourceHttpError(t *testing.T) {
 	var data = []byte(`test contents`)
 	var fs = afero.NewMemMapFs()
 	require.NoError(t, afero.WriteFile(fs, "/github.com/something", data, 0644))
-	_, err := readSource("http://github.com/something", "/c",
+	_, err := ReadSource("http://github.com/something", "/c",
 		map[string]afero.Fs{"file": afero.NewMemMapFs(), "https": fs}, nil)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), `only supported schemes for imports are file and https`)
